+++ NEWS.md
@@ -1,3 +1,5 @@
+* Do not use arrow in unit tests
+
 # MolgenisArmadillo 1.0.0
 
 * Release to CRAN